# unpack: optimize unpack.Tgz memory allocations