In the worst case (when the `io.Reader` isnt an `io.ReadSeeker`), `unpack.Tgz` would use `io.ReadAll` on the `io.Reader` in order to fall-back to `unpack.Tar` if the file extension was incorrect (.tar.gz for something what wasnt actually gzipped).  

After https://github.com/sourcegraph/sourcegraph/pull/44497/, https://github.com/sourcegraph/sourcegraph/pull/44517/ and https://github.com/sourcegraph/sourcegraph/pull/44520, this worst will always be triggered, as http response bodies are not `io.ReadSeeker`s. This can cause excessive memory copying from `io.ReadAll`, where we could be reading a minimal amount to determine whether the file is actually gzipped or not.

Benchmark figures, where Old is the original, and New is the new (of course):
```
$ go test -benchmem -run='^$' -bench '^BenchmarkTgzFallback$' github.com/sourcegraph/sourcegraph/internal/unpack -v -count=10 -race
goos: linux
goarch: amd64
pkg: github.com/sourcegraph/sourcegraph/internal/unpack
cpu: AMD Ryzen 5 5600X 6-Core Processor             
BenchmarkTgzFallback
BenchmarkTgzFallback/Old
BenchmarkTgzFallback/Old-12                29144             43833 ns/op           14557 B/op         36 allocs/op
BenchmarkTgzFallback/Old-12                26395             45004 ns/op           14574 B/op         36 allocs/op
BenchmarkTgzFallback/Old-12                28130             45917 ns/op           14599 B/op         36 allocs/op
BenchmarkTgzFallback/Old-12                26758             46788 ns/op           14564 B/op         36 allocs/op
BenchmarkTgzFallback/Old-12                25844             45694 ns/op           14532 B/op         36 allocs/op
BenchmarkTgzFallback/Old-12                27319             45318 ns/op           14512 B/op         36 allocs/op
BenchmarkTgzFallback/Old-12                27026             46731 ns/op           14586 B/op         36 allocs/op
BenchmarkTgzFallback/Old-12                26932             47716 ns/op           14538 B/op         36 allocs/op
BenchmarkTgzFallback/Old-12                25428             45836 ns/op           14563 B/op         36 allocs/op
BenchmarkTgzFallback/Old-12                26865             41620 ns/op           14530 B/op         36 allocs/op
BenchmarkTgzFallback/New
BenchmarkTgzFallback/New-12                33439             40709 ns/op            5945 B/op         32 allocs/op
BenchmarkTgzFallback/New-12                29628             39881 ns/op            5940 B/op         31 allocs/op
BenchmarkTgzFallback/New-12                31644             39578 ns/op            5971 B/op         32 allocs/op
BenchmarkTgzFallback/New-12                29116             39691 ns/op            5957 B/op         32 allocs/op
BenchmarkTgzFallback/New-12                30648             39306 ns/op            5995 B/op         32 allocs/op
BenchmarkTgzFallback/New-12                31161             41873 ns/op            5984 B/op         32 allocs/op
BenchmarkTgzFallback/New-12                30362             41084 ns/op            5939 B/op         32 allocs/op
BenchmarkTgzFallback/New-12                30147             39431 ns/op            5999 B/op         32 allocs/op
BenchmarkTgzFallback/New-12                31102             39300 ns/op            5933 B/op         32 allocs/op
BenchmarkTgzFallback/New-12                32269             42896 ns/op            5977 B/op         32 allocs/op
```

## Test plan

Unit test, benchmark, code review, live :)